+++ FrontEnd/src/pages/vocabularies/index.js
@@ -14,11 +14,10 @@ import TablePagination from '@mui/material/TablePagination';
 import TableRow from '@mui/material/TableRow';
 import Paper from '@mui/material/Paper';
 import IconButton from '@mui/material/IconButton';
-import FirstPageIcon from '@mui/icons-material/FirstPage';
+import FirstPage from '@mui/icons-material/FirstPage';
 import KeyboardArrowLeft from '@mui/icons-material/KeyboardArrowLeft';
 import KeyboardArrowRight from '@mui/icons-material/KeyboardArrowRight';
-import LastPageIcon from '@mui/icons-material/LastPage';
-
+import LastPage from '@mui/icons-material/LastPage';
 
 function TablePaginationActions(props) {
   const theme = useTheme();
@@ -43,7 +42,7 @@ function TablePaginationActions(props) {
             disabled={page === 0}
             aria-label="first page"
         >
-            {theme.direction === 'rtl' ? <LastPageIcon /> : <FirstPageIcon />}
+            {theme.direction === 'rtl' ? <LastPage /> : <FirstPage />}
         </IconButton>
         <IconButton
             onClick={handleBackButtonClick}
@@ -64,7 +63,7 @@ function TablePaginationActions(props) {
             disabled={page >= Math.ceil(count / rowsPerPage) - 1}
             aria-label="last page"
         >
-            {theme.direction === 'rtl' ? <FirstPageIcon /> : <LastPageIcon />}
+            {theme.direction === 'rtl' ? <FirstPage /> : <LastPage />}
         </IconButton>
     </Box>
   );
