+++ FrontEnd/src/redux/reducers/index.js
@@ -1,8 +1,7 @@
 import { combineReducers } from 'redux';
 import Categories from './categories'
 import Posts from './posts'
-import Account from './account';
 
 export default combineReducers({
-  Categories, Posts, Account
+  Categories, Posts,
 });
